requirements.txt
# Flask Web Framework
flask>=3.0.0
python-dotenv>=1.0.0
werkzeug>=3.0.0

# OpenAI API
openai>=1.30.0

# Document Processing
PyPDF2>=3.0.1
python-docx>=1.1.0

# Vector Search and ML
numpy>=1.26.0
scikit-learn>=1.3.0

# Text Processing
tiktoken>=0.5.1